+++ test/event/handle_event_test.exs
@@ -11,9 +11,15 @@ defmodule Commanded.Event.HandleEventTest do
   alias Commanded.ExampleDomain.AccountBalanceHandler
   alias Commanded.ExampleDomain.BankAccount.Events.{BankAccountOpened,MoneyDeposited}
 
+  setup do
+    on_exit fn ->
+      Commanded.Helpers.Process.shutdown(AccountBalanceHandler)
+      Commanded.Helpers.Process.shutdown(AppendingEventHandler)
+    end
+  end
+
   test "should be notified of events" do
-    {:ok, _} = AccountBalanceHandler.start_link()
-    {:ok, handler} = Commanded.Event.Handler.start_link("account_balance", AccountBalanceHandler)
+    {:ok, handler} = AccountBalanceHandler.start_link()
 
     events = [
       %BankAccountOpened{account_number: "ACC123", initial_balance: 1_000},
@@ -31,8 +37,7 @@ defmodule Commanded.Event.HandleEventTest do
   defmodule UninterestingEvent, do: defstruct [field: nil]
 
   test "should ignore uninterested events" do
-    {:ok, _} = AccountBalanceHandler.start_link
-		{:ok, handler} = Commanded.Event.Handler.start_link("account_balance", AccountBalanceHandler)
+    {:ok, handler} = AccountBalanceHandler.start_link
 
     # include uninterested events within those the handler is interested in
     events = [
@@ -51,9 +56,7 @@ defmodule Commanded.Event.HandleEventTest do
     end)
   end
 
-  test "should ignore events created before the event handler's subscription when starting from `current`" do
-    {:ok, _} = AppendingEventHandler.start_link()
-
+  test "should ignore events created before the event handler's subscription when starting from `:current`" do
     stream_uuid = UUID.uuid4
     initial_events = [%BankAccountOpened{account_number: "ACC123", initial_balance: 1_000}]
     new_events = [%MoneyDeposited{amount: 50, balance: 1_050}]
@@ -62,7 +65,7 @@ defmodule Commanded.Event.HandleEventTest do
 
     wait_for_event BankAccountOpened
 
-    {:ok, _handler} = Commanded.Event.Handler.start_link("test_event_handler", AppendingEventHandler, start_from: :current)
+    {:ok, _} = AppendingEventHandler.start_link(start_from: :current)
 
     {:ok, 2} = EventStore.append_to_stream(stream_uuid, 1, Commanded.Event.Mapper.map_to_event_data(new_events, UUID.uuid4))
 
@@ -70,20 +73,24 @@ defmodule Commanded.Event.HandleEventTest do
 
     Wait.until(fn ->
       assert AppendingEventHandler.received_events == new_events
-      assert pluck(AppendingEventHandler.received_metadata, :stream_version) == [2]
+
+      [ metadata ] = AppendingEventHandler.received_metadata()
+
+      assert Map.get(metadata, :event_number) == 2
+      assert Map.get(metadata, :stream_id) == stream_uuid
+      assert Map.get(metadata, :stream_version) == 2
+      assert %NaiveDateTime{} = Map.get(metadata, :created_at)
     end)
 	end
 
-  test "should receive events created before the event handler's subscription when starting from `origin`" do
-    {:ok, _} = AppendingEventHandler.start_link()
-
+  test "should receive events created before the event handler's subscription when starting from `:origin`" do
     stream_uuid = UUID.uuid4
     initial_events = [%BankAccountOpened{account_number: "ACC123", initial_balance: 1_000}]
     new_events = [%MoneyDeposited{amount: 50, balance: 1_050}]
 
     {:ok, 1} = EventStore.append_to_stream(stream_uuid, 0, Commanded.Event.Mapper.map_to_event_data(initial_events, UUID.uuid4))
 
-    {:ok, _handler} = Commanded.Event.Handler.start_link("test_event_handler", AppendingEventHandler, start_from: :origin)
+    {:ok, _} = AppendingEventHandler.start_link(start_from: :origin)
 
     {:ok, 2} = EventStore.append_to_stream(stream_uuid, 1, Commanded.Event.Mapper.map_to_event_data(new_events, UUID.uuid4))
 
@@ -91,13 +98,21 @@ defmodule Commanded.Event.HandleEventTest do
 
     Wait.until(fn ->
       assert AppendingEventHandler.received_events == initial_events ++ new_events
-      assert pluck(AppendingEventHandler.received_metadata, :stream_version) == [1, 2]
+
+      received_metadata = AppendingEventHandler.received_metadata()
+
+      assert pluck(received_metadata, :event_number) == [1, 2]
+      assert pluck(received_metadata, :stream_version) == [1, 2]
+
+      Enum.each(received_metadata, fn metadata ->
+        assert Map.get(metadata, :stream_id) == stream_uuid
+        assert %NaiveDateTime{} = Map.get(metadata, :created_at)
+      end)
     end)
 	end
 
 	test "should ignore already seen events" do
-    {:ok, _} = AppendingEventHandler.start_link
-    {:ok, handler} = Commanded.Event.Handler.start_link("test_event_handler", AppendingEventHandler)
+    {:ok, handler} = AppendingEventHandler.start_link()
 
     events = [
       %BankAccountOpened{account_number: "ACC123", initial_balance: 1_000},